requirements.txt
colorama==0.4.6
pillow==11.1.0
pypng==0.20220715.0
PyQRCode==1.2.1
